🚧 WIP: add dymension widgets
